The intends-to-provide code wasn't working quite right.  If I #include <vector>, then I intend-to-provide <vector>: you don't have to #include <vector> just because I use it in my (presumably templated) code, because I'm #including <vector> for you.  However, this logic runs before the private->public mappings are done, so you're never actually wondering about <vector>, you're wondering about <bits/stl_vector.h>.
In other words, if I'm #including <vector>, it's not to get
the symbols just provided by <vector>, but to get the symbols
provided by all the private headers that map to <vector> as
well.

I've changed the code that populates the intends-to-provide
map to handle this case.  Now we have many fewer instances of
iwyu randomly deciding that you need to #include a file that
is only used in templated code.  I think there are more
improvements to make in this area, but that will have to wait
for another day.

Another problem is that when we decided a template was
responsible for a decl, because it intends-to-provide it, we
were ignoring the decl, rather than keeping it and just
assigning its use to the template.  Usually this is harmless,
but it could result in us deciding that the decl isn't used
anywhere at all, and removing an #include (from the template
file) that should be kept.  In addition to fixing the relevant
code, I've updated a test to make sure that doesn't happen.

Testing showed up another issue, which is that the code for
determining decl locations was correctly handling templated
classes but not templated functions.  This is now fixed.
Fixing it, unfortunately, made visible a known weakness in
IntendsToProvide, involving its ability to guess whether a
file actually intends to provide a definition for its symbols
or not, which is now a new TODO in badinc.cc.

R=dsturtevant
DELTA=285  (196 added, 66 deleted, 23 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=736


git-svn-id: http://include-what-you-use.googlecode.com/svn/trunk@48 812d1f90-f944-d88a-411a-122bdc93e1ea