The latest clang update changed the way elaborated types were handled, which broke iwyu in several different ways.  We didn't notice at the time because run_iwyu_tests was silently broken and not running any tests.  Now that that's fixed, we see the trouble.
This CL fixes a few of the problems -- more remain that I
haven't been able to figure out yet.  Basically, an elaborated
class type (a superclass in something like Foo::FooSubclass)
is now stored as an elaboratedtype that's a sibling to
FooSubclass, rather than as a NestedNameSpecifier that's a
parent to FooSubclass.  I'm not sure I understand why, but
that's not my job...

The main fix is to add a new IsClassElaborationQualifier()
function, which abstracts out the test for whether we're a
superclass in a context like Foo::FooSubclass.  We now use
that both for regular types and for template-argument types.

Another, seemingly unrelated fix -- that may be due to a
different change in clang -- fixes the handling of
typedefs-to-typedefs.  I don't know why they worked before,
actually, so maybe this was a bugfix in clang.

More or less unrelatedly, also do a few style cleanups in the
preprocessor file.

R=dsturtevant
DELTA=51  (39 added, 2 deleted, 10 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=892


git-svn-id: http://include-what-you-use.googlecode.com/svn/trunk@67 812d1f90-f944-d88a-411a-122bdc93e1ea