Handle dependencies on 'using' declarations properly.  If we use a type via a using declaration, we need to make sure to use that using declaration as well.  It may be in a different file from the type!
There may be corner cases we don't handle right yet, but we do
the right thing at least most of the time.  One case we don't
handle is when there's a using declaration but we ignore it
('using foo::a; foo::a = 1').  We'd do better if clang would
report whether a given decl-use was via a using declaration or
not, but right now it doesn't.  cf:
http://lists.cs.uiuc.edu/pipermail/cfe-dev/2011-April/014564.html
(though that applies only to using vars, not functions).

In order to implement this reasonably, I had to (finally!)
unify the type and decl reporting in iwyu.cc, rather than
iwyu_output.cc.  This allows me to remove a bit of duplicated
code, and also makes the logic easier to follow.

R=dsturtevant
DELTA=358  (286 added, 71 deleted, 1 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=1576


git-svn-id: http://include-what-you-use.googlecode.com/svn/trunk@150 812d1f90-f944-d88a-411a-122bdc93e1ea