Add special support for third-party code, to make it less likely we'll suggest adding an internal third-party header. Basically, we stop trying to do include-what-you-use fixes on third-party code.
More precisely, we make an 'implicit' judgment on which
third-party headers are public and which are private, based on
what existing code (in this translation unit) #includes.  We
marked all unincluded files as private, which means iwyu will
never suggest adding a new third-party file as an #include.
Insted, it will suggest some already-included third-party file
that gets the needed file transitively.

Since it's not really practical for us to fix third-party code
to have better #include hygiene, or even to mark up third-party
code with iwyu pragmas, we need to do something similar to
this.  We could just manually update iwyu_include_picker's
third_party_map with rules for every third party package we
have, but that's expensive.  This is much cheaper, with the
downside that we may miss some potential include-what-you-use
opportunities in third-party code (going from a more-generic
third-party include to a less-generic one).  I think that's a
low cost.

R=dsturtevant
DELTA=113  (104 added, 0 deleted, 9 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=1577


git-svn-id: http://include-what-you-use.googlecode.com/svn/trunk@151 812d1f90-f944-d88a-411a-122bdc93e1ea