First shot at allowing code-authors to use forward-declares as a signal to iwyu that it should not use its stricter-than-required rules for typedefs and autocasts (that is, iwyu requires the full type for Foo in 'typedef Foo MyTypedef' while C++ does not).
It's a 'first shot' because we make no attempt to deal with
template arguments that may appear in underlying types.  But
otherwise, it should be pretty complete.

Along the way -- not entirely related, but showed up in
testing -- fixed a place where we were declaring a use where
we shouldn't have been: when forward-declaring a template
type, we were were saying we 'used' the type we were
declaring.  This isn't true: a declaration isn't a use.
Normally it didn't cause any problems (the declaration and use
are in the same file!) but it did in the test, because it was
a forward-decalred explicit specialization, which is I guess
something I'd never done before.

R=wan,dsturtevant
DELTA=628  (567 added, 18 deleted, 43 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=1602


git-svn-id: http://include-what-you-use.googlecode.com/svn/trunk@176 812d1f90-f944-d88a-411a-122bdc93e1ea