When doing some recursive calls while reporting type use, we were making virtual function calls that were not appropriate. It got to be a problem when the virtual function call (ReportTypeUse) was doing type-resugaring and getting back to the same original type that we were trying to recurse on! That is, TypedefType -> recurse on UnderlyingType -> resugared to TypedefType -> recurse on UnderlyingType -> infinite loop.
I believe this is the only bad situation we can get in in
practice, since the other recursive calls are all on decls
which don't have a problem with resugaring.  But the whole
thing is really fragile.  The right way to fix it, though, is
to separate out the template-instantiation code from the
iwyu-visiting code, so I'm just doing band-aids for now.
Maybe I'm recovered enough from the fixit to try the right fix
(which is quite invasive) sometime soon.

R=dsturtevant
DELTA=74  (72 added, 0 deleted, 2 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=2189


git-svn-id: http://include-what-you-use.googlecode.com/svn/trunk@259 812d1f90-f944-d88a-411a-122bdc93e1ea