We already ignored cycles-in-includes in third-party files. Ignore them for files in /internal/ as well, which suffer from the same potential problem.
We could decide to complain on such cycles in /internal/ as
well, since this is our code so we can fix these cyclic
includes.  But that requires adding new infrastructure, to add
a new kind of iwyu complaint, and I don't think this is
important enough to warrant.  I'd rather the cycle get broken
just by the normal include cleanup that iwyu does.

Implementation-wise, I would have preferred to add a bit to
each include-mapping we see, to indicate whether it's one we
should complain if we see a cycle, or not.  But that was a
pretty invasive change to the data structures that are complex
enough already, so I didn't.  I think this code is already
hard enough to understand without adding another layer of
abstraction necessary to make such a bit easy to access.  I
could be convinced otherwise.

R=dsturtevant
DELTA=65  (63 added, 0 deleted, 2 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=2651


git-svn-id: http://include-what-you-use.googlecode.com/svn/trunk@275 812d1f90-f944-d88a-411a-122bdc93e1ea