CanIgnoreCurrentASTNode() is an important optimization to avoid doing too much work.  But we have to be careful not to ignore stuff that can actually affect the current translation unit.
We found an instance where that was happening.  The problem
was macros: We have to be careful not to ignore files that
define macros we use.  That is because oten we attribute
stuff in macros to the file that macro is written in, but the
code can directly affect code where the macro is expanded.
Thus, we should consider both the macro's as-written location
and its expansion location when deciding whether it's safe to
ignore.

Note we're being a bit conservative here, but since
CanIgnoreCurrentASTNode() is purely an optimization, that's
the right thing to do.

R=dsturtevant
DELTA=26  (24 added, 0 deleted, 2 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=3017


git-svn-id: http://include-what-you-use.googlecode.com/svn/trunk@295 812d1f90-f944-d88a-411a-122bdc93e1ea