Merge from trunk to clang_3.5 branch r564: ------------------------------------------------------------------------ r564 | kim.grasman@gmail.com | 2014-08-18 20:45:03 +0200 (Mon, 18 Aug 2014) | 9 lines
Fix one of the bugs from issue 144: confusion with folders

The heuristic used to find the associated header of the main file
involved checking the name of the first file included.
When --pch_in_code is provided, the first file included is _always_
the precompiled header and can't be the associated header.
Have the heuristic check the second header with --pch_in_code.

Based on patch by David Punset, thanks!
------------------------------------------------------------------------

git-svn-id: http://include-what-you-use.googlecode.com/svn/branches/clang_3.5@575 812d1f90-f944-d88a-411a-122bdc93e1ea