Rename GetFileEntry(FullSourceLoc) GetLocFileEntry
The use of FullSourceLoc in IWYU was confusing and error prone.
GetFileEntry was already templated on all AST node types and overloaded on
SourceLocation, so calling GetFileEntry with a FullSourceLoc would have a
completely different meaning than calling it with a plain SourceLocation.

Add comments to make it clearer that GetLocFileEntry does not do any magic.

Also inline a few pointless utility methods.

No functional change.

git-svn-id: http://include-what-you-use.googlecode.com/svn/trunk@604 812d1f90-f944-d88a-411a-122bdc93e1ea